+++ flutter/test/mocks.mocks.dart
@@ -563,7 +563,6 @@ class MockSentryTracer extends _i1.Mock implements _i3.SentryTracer {
 /// A class which mocks [SentryTransaction].
 ///
 /// See the documentation for Mockito's code generation for more information.
-// ignore: must_be_immutable
 class MockSentryTransaction extends _i1.Mock implements _i2.SentryTransaction {
   MockSentryTransaction() {
     _i1.throwOnMissingStub(this);
@@ -639,10 +638,160 @@ class MockSentryTransaction extends _i1.Mock implements _i2.SentryTransaction {
       ) as _i2.SentryId);
 
   @override
+  set eventId(_i2.SentryId? _eventId) => super.noSuchMethod(
+        Invocation.setter(#eventId, _eventId),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set timestamp(DateTime? _timestamp) => super.noSuchMethod(
+        Invocation.setter(#timestamp, _timestamp),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set platform(String? _platform) => super.noSuchMethod(
+        Invocation.setter(#platform, _platform),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set logger(String? _logger) => super.noSuchMethod(
+        Invocation.setter(#logger, _logger),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set serverName(String? _serverName) => super.noSuchMethod(
+        Invocation.setter(#serverName, _serverName),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set release(String? _release) => super.noSuchMethod(
+        Invocation.setter(#release, _release),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set dist(String? _dist) => super.noSuchMethod(
+        Invocation.setter(#dist, _dist),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set environment(String? _environment) => super.noSuchMethod(
+        Invocation.setter(#environment, _environment),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set modules(Map<String, String>? _modules) => super.noSuchMethod(
+        Invocation.setter(#modules, _modules),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set message(_i2.SentryMessage? _message) => super.noSuchMethod(
+        Invocation.setter(#message, _message),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set exceptions(List<_i2.SentryException>? _exceptions) => super.noSuchMethod(
+        Invocation.setter(#exceptions, _exceptions),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set threads(List<_i2.SentryThread>? _threads) => super.noSuchMethod(
+        Invocation.setter(#threads, _threads),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set transaction(String? _transaction) => super.noSuchMethod(
+        Invocation.setter(#transaction, _transaction),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set level(_i2.SentryLevel? _level) => super.noSuchMethod(
+        Invocation.setter(#level, _level),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set culprit(String? _culprit) => super.noSuchMethod(
+        Invocation.setter(#culprit, _culprit),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set tags(Map<String, String>? _tags) => super.noSuchMethod(
+        Invocation.setter(#tags, _tags),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set extra(Map<String, dynamic>? _extra) => super.noSuchMethod(
+        Invocation.setter(#extra, _extra),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set breadcrumbs(List<_i2.Breadcrumb>? _breadcrumbs) => super.noSuchMethod(
+        Invocation.setter(#breadcrumbs, _breadcrumbs),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set user(_i2.SentryUser? _user) => super.noSuchMethod(
+        Invocation.setter(#user, _user),
+        returnValueForMissingStub: null,
+      );
+
+  @override
   _i2.Contexts get contexts => (super.noSuchMethod(
         Invocation.getter(#contexts),
         returnValue: _FakeContexts_6(this, Invocation.getter(#contexts)),
       ) as _i2.Contexts);
+
+  @override
+  set contexts(_i2.Contexts? _contexts) => super.noSuchMethod(
+        Invocation.setter(#contexts, _contexts),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set fingerprint(List<String>? _fingerprint) => super.noSuchMethod(
+        Invocation.setter(#fingerprint, _fingerprint),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set sdk(_i2.SdkVersion? _sdk) => super.noSuchMethod(
+        Invocation.setter(#sdk, _sdk),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set request(_i2.SentryRequest? _request) => super.noSuchMethod(
+        Invocation.setter(#request, _request),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set debugMeta(_i2.DebugMeta? _debugMeta) => super.noSuchMethod(
+        Invocation.setter(#debugMeta, _debugMeta),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  set type(String? _type) => super.noSuchMethod(
+        Invocation.setter(#type, _type),
+        returnValueForMissingStub: null,
+      );
 
   @override
   Map<String, dynamic> toJson() => (super.noSuchMethod(
@@ -1281,6 +1430,14 @@ class MockSentryNativeBinding extends _i1.Mock
           ),
         ),
       ) as _i11.FutureOr<_i2.SentryId>);
+
+  @override
+  _i11.FutureOr<void> startSession({bool? ignoreDuration = false}) =>
+      (super.noSuchMethod(
+        Invocation.method(#startSession, [], {
+          #ignoreDuration: ignoreDuration,
+        }),
+      ) as _i11.FutureOr<void>);
 }
 
 /// A class which mocks [SentryDelayedFramesTracker].
@@ -2023,12 +2180,16 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   int scheduleFrameCallback(
     _i21.FrameCallback? callback, {
     bool? rescheduling = false,
+    bool? scheduleNewFrame = true,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
           #scheduleFrameCallback,
           [callback],
-          {#rescheduling: rescheduling},
+          {
+            #rescheduling: rescheduling,
+            #scheduleNewFrame: scheduleNewFrame,
+          },
         ),
         returnValue: 0,
       ) as int);
@@ -2307,6 +2468,24 @@ class MockWidgetsFlutterBinding extends _i1.Mock
       );
 
   @override
+  void addSemanticsActionListener(
+    _i8.ValueSetter<_i6.SemanticsActionEvent>? listener,
+  ) =>
+      super.noSuchMethod(
+        Invocation.method(#addSemanticsActionListener, [listener]),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  void removeSemanticsActionListener(
+    _i8.ValueSetter<_i6.SemanticsActionEvent>? listener,
+  ) =>
+      super.noSuchMethod(
+        Invocation.method(#removeSemanticsActionListener, [listener]),
+        returnValueForMissingStub: null,
+      );
+
+  @override
   _i12.SemanticsHandle ensureSemantics() => (super.noSuchMethod(
         Invocation.method(#ensureSemantics, []),
         returnValue: _FakeSemanticsHandle_32(
@@ -2550,6 +2729,25 @@ class MockSentryJsBinding extends _i1.Mock implements _i23.SentryJsBinding {
   @override
   void captureEnvelope(List<Object>? envelope) => super.noSuchMethod(
         Invocation.method(#captureEnvelope, [envelope]),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  void startSession() => super.noSuchMethod(
+        Invocation.method(#startSession, []),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  void updateSession({int? errors, String? status}) => super.noSuchMethod(
+        Invocation.method(
+            #updateSession, [], {#errors: errors, #status: status}),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  void captureSession() => super.noSuchMethod(
+        Invocation.method(#captureSession, []),
         returnValueForMissingStub: null,
       );
 }